API

Coloring

QuasiStableColors.refine_fixpoint — Method

refine_fixpoint(
    G::AbstractGraph{T},
    weights = nothing,
    special =  Set{T}(),
    warm_start = Vector{Vector{T}}(),
    early_stop = Inf,
    eps = 0.0,
)

Compute a quasi-stable coloring for the undirected graph G. Typically, you should set one of:

eps: maximum q-error allowed

early_stop: number of colors to use

Optional parameters:

warm_start: coloring to refine. If not provided, start using trivial (single color) partitioning assumed.

QuasiStableColors.refine_stable — Method

refine_stable(G::AbstractGraph{T})

Compute the stable coloring for the undirected graph G. Provided for comparasion.

QuasiStableColors.refine_bipartite — Function

Equivalent to refine_fixpoint but optimized for bipartite graphs. Faster but less general.

Maximum-flow

QuasiStableColors.Flow.lifted_maxflow — Function

Compute the maximum flow from s to t in flow network G. Capacities defined using weights=.

Betweenness centrality

QuasiStableColors.Centrality.approx_betweenness_centrality — Function

Approximate betweenness centrality using a size colors q-stable coloring.

Linear programming

QuasiStableColors.Optimize.lifted_minimize — Function

Approximate linear program min c^T x where A x >= b, x >=0.